+++ Sources/Base/OAuth2Error.swift
@@ -67,6 +67,9 @@ public enum OAuth2Error: Error, CustomStringConvertible, Equatable {
 	/// The redirect URL is invalid; with explanation.
 	case invalidRedirectURL(String)
 	
+	/// There is no access token.
+	case noAccessToken
+	
 	/// There is no refresh token.
 	case noRefreshToken
 	
@@ -205,6 +208,8 @@ public enum OAuth2Error: Error, CustomStringConvertible, Equatable {
 			return "Invalid authorization context"
 		case .invalidRedirectURL(let url):
 			return "Invalid redirect URL: \(url)"
+		case .noAccessToken:
+			return "I don't have an access token, cannot sign request"
 		case .noRefreshToken:
 			return "I don't have a refresh token, not trying to refresh"
 		
@@ -239,7 +244,7 @@ public enum OAuth2Error: Error, CustomStringConvertible, Equatable {
 			return "Failed to decode given data as a UTF-8 string"
 		
 		case .unauthorizedClient:
-			return "The client is not authorized to request an access token using this method."
+			return "Unauthorized"
 		case .forbidden:
 			return "Forbidden"
 		case .wrongUsernamePassword:
@@ -277,6 +282,7 @@ public enum OAuth2Error: Error, CustomStringConvertible, Equatable {
 		case (.noAuthorizationContext, .noAuthorizationContext):                 return true
 		case (.invalidAuthorizationContext, .invalidAuthorizationContext):       return true
 		case (.invalidRedirectURL(let lhu), .invalidRedirectURL(let rhu)):       return lhu == rhu
+		case (.noAccessToken, .noAccessToken):                       return true
 		case (.noRefreshToken, .noRefreshToken):			         return true
 		
 		case (.notUsingTLS, .notUsingTLS):                           return true