+++ Sources/DataLoader/OAuth2DataLoader.swift
@@ -37,9 +37,20 @@ open class OAuth2DataLoader: OAuth2Requestable {
 	public var alsoIntercept403: Bool = false
 	
 	
-	public init(oauth2: OAuth2) {
+	/**
+	Designated initializer.
+	
+	Provide `host` if 301 and 302 redirects should be followed automatically, as long as they appear on the same host.
+	
+	- parameter oauth2: The OAuth2 instance to use for authorization when loading data.
+	- parameter host:   If given will handle redirects within the same host by way of `OAuth2DataLoaderSessionTaskDelegate`
+	*/
+	public init(oauth2: OAuth2, host: String? = nil) {
 		self.oauth2 = oauth2
 		super.init(logger: oauth2.logger)
+		if let host = host {
+			sessionDelegate = OAuth2DataLoaderSessionTaskDelegate(loader: self, host: host)
+		}
 	}
 	
 	
@@ -129,7 +140,7 @@ open class OAuth2DataLoader: OAuth2Requestable {
 					callback(response)
 				}
 			}
-				
+			
 			// some other error, pass along
 			catch {
 				callback(response)
@@ -138,7 +149,7 @@ open class OAuth2DataLoader: OAuth2Requestable {
 	}
 	
 	/**
-	If not already authorizing, will use its `oauth2` instance to start authorization.
+	If not already authorizing, will use its `oauth2` instance to start authorization after forgetting any tokens (!).
 	
 	This method will ignore calls while authorization is ongoing, meaning you will only get the callback once per authorization cycle.
 	
@@ -148,6 +159,7 @@ open class OAuth2DataLoader: OAuth2Requestable {
 	open func attemptToAuthorize(callback: @escaping ((OAuth2JSON?, OAuth2Error?) -> Void)) {
 		if !isAuthorizing {
 			isAuthorizing = true
+			oauth2.forgetTokens()
 			oauth2.authorize() { authParams, error in
 				self.isAuthorizing = false
 				callback(authParams, error)
@@ -201,8 +213,13 @@ open class OAuth2DataLoader: OAuth2Requestable {
 	func retryAll() {
 		dequeueAndApply() { req in
 			var request = req.request
-			request.sign(with: oauth2)
-			self.perform(request: request, retry: false, callback: req.callback)
+			do {
+				try request.sign(with: oauth2)
+				self.perform(request: request, retry: false, callback: req.callback)
+			}
+			catch let error {
+				NSLog("OAuth2.DataLoader.retryAll(): \(error)")
+			}
 		}
 	}